Update zfs_snapshot_replication.sh
Add to only run after array is started. If this script is scheduled, and your server has been off for an extended period of time, it could try to run this script while the array is starting. This will hose everything up and prevent the zfs drives from being mounted and prevent the array from starting.